[Behat] Extract Zone transformer
