+++ src/Sylius/Behat/spec/Context/Setup/ZoneContextSpec.php
@@ -73,49 +73,12 @@ class ZoneContextSpec extends ObjectBehavior
         $this->thereIsRestOfTheWorldZoneContainingAllOtherCountries();
     }
 
-    function it_sets_default_zone($zoneRepository, $settingsManager, Settings $settings, ZoneInterface $zone)
+    function it_sets_default_zone($settingsManager, Settings $settings, ZoneInterface $zone)
     {
-        $zoneRepository->findOneBy(['code' => 'EU'])->willReturn($zone);
-
         $settingsManager->loadSettings('sylius_taxation')->willReturn($settings);
         $settings->set('default_tax_zone', $zone)->shouldBeCalled();
         $settingsManager->saveSettings('sylius_taxation', $settings)->shouldBeCalled();
 
-        $this->defaultTaxZoneIs('EU');
-    }
-
-    function it_throws_exception_if_zone_with_given_code_does_not_exist_while_setting_default_zone($zoneRepository)
-    {
-        $zoneRepository->findOneBy(['code' => 'EU'])->willReturn(null);
-
-        $this->shouldThrow(new \InvalidArgumentException('Zone with code "EU" does not exist.'))->during('defaultTaxZoneIs', ['EU']);
-    }
-
-    function it_returns_zone_by_its_code($zoneRepository, ZoneInterface $zone)
-    {
-        $zoneRepository->findOneBy(['code' => 'EU'])->willReturn($zone);
-
-        $this->getZoneByCode('EU')->shouldReturn($zone);
-    }
-
-    function it_throws_exception_if_zone_with_given_code_does_not_exist($zoneRepository)
-    {
-        $zoneRepository->findOneBy(['code' => 'EU'])->willReturn(null);
-
-        $this->shouldThrow(new \Exception('Zone with code "EU" does not exist.'))->during('getZoneByCode', ['EU']);
-    }
-
-    function it_returns_the_rest_of_the_world_zone($zoneRepository, ZoneInterface $zone)
-    {
-        $zoneRepository->findOneBy(['code' => 'RoW'])->willReturn($zone);
-
-        $this->getRestOfTheWorldZone()->shouldReturn($zone);
-    }
-
-    function it_throws_exception_if_there_is_no_rest_of_the_world_zone($zoneRepository)
-    {
-        $zoneRepository->findOneBy(['code' => 'RoW'])->willReturn(null);
-
-        $this->shouldThrow(new \Exception('Rest of the world zone does not exist.'))->during('getRestOfTheWorldZone');
+        $this->defaultTaxZoneIs($zone);
     }
 }
